requirements.txt
adbutils==1.2.12
aiolimiter==1.1.0
beautifulsoup4==4.12.3
cairosvg==2.7.1
colorama==0.4.6
gymnasium==0.28.1
html2text==2020.1.16
langchain==0.1.4
langchain_core==0.1.17
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
openai==0.27.6
pandas==2.0.2
pydantic==1.10.9
pygal==3.0.4
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.31.0
seaborn==0.13.2
spacy==3.7.2
streamlit==1.27.2
tiktoken==0.5.1
tqdm==4.65.0
trafilatura==1.6.1
transformers==4.29.0
uiautomator2
xmltodict==0.13.0
